Revamp cleanup
Firstly, remove the too-optimistic use of YAML anchors in the values
file, they're not properly replaced when a user specifies a partial yaml
file to override some values.

Also, move the cleanup template to use two sections of hooks
pre-delete, which:
 - delete the storageoscluster CR
 - wait for it's pods to come down (to prevent the operator being
   deleted before it can do its job vis-a-vis resource deletion)
 and post-delete, which
 - delete 4 misc operator resources (CMs & secrets) that were being
   left behind after operator deletion
Then let helm delete all the resources that it created.